Turn the fibonacci program into a test
The fibonacci program was a test, and it also includes a test run of
printf features, so let's turn into a test. Build it for 'virt' machine,
capture its output, and compare it to a reference output.

'make test' will print 'OK' on success, or a diff if something breaks.