Fix array out of bounds write
Since curr_proc is init'ed to -1, the first call to kernel_timer_tick
used to write to proc_table[-1]. Which was a hilarious bug for bug
compatibility, since we needed to discard the first mepc in order to
work correctly.

This bug did not overwrite nothing important yet, but boy, would it do
so later, in a hilarious backstab!