Fix warning in kprintp
Unhardcode the expected fixed size of pointer, make it dependent on the
actual sizeof(p).

Also, fix a bug with stack handling in kprints and sys_puts. I was
misusing the sx/lx macros by specifying an offset in bytes to them instead
of an offset in registers. Use stackalloc_x/stackfree_x along with sx/lx
to symmetrically specify number of registers to both.