Rewrite hart sync code during early boot
We used to use an atomic op, but that's unreliable while memory segments
are not init'ed: works fine on qemu, but will not work on physical
hardware.

This new approach avoids using main memory for synchronizing harts. It
hardcodes the 0th hart to be responsible for init'ing the memory
segments. Other harts spin a busy loop while the boot hart is done with
it. The harts communicate via MSIP register, which allows reliable
signalling between harts.

The approach is taken from the Freedom SDK CRT code.