Make it run on HiFive1 Rev B
There were two major problems with our code on HiFive: wrong trap_vector
and wrong syscall_vector.

Trap vector had two problems of its own: first, on HiFive a vectored
trap vectored is required to be aligned to 64 bytes. Second, with
compact instructions the vector was too dense, so we need to align each
entry to 4 bytes as well. Did the same for exception_vector as well.

The problem with syscall vector was different: the global variable got
placed in a .data segment by default, as it should be. But we're still
missing code that copies .data segment from ROM to RAM, so in effect the
vector was garbage. As a quick fix I moved syscall_vector to .text, as
it's meant to be read-only anyway.

Finally, a couple minor fixes: moved the clock freq definition to
machine-specific headers, as they differ on QEMU and HiFive. Also,
commented out some noisy stuff from userland code. I will factor it out
to some dedicated test later.