Reenable integration tests
The problem was the new GCC version building with Global Offset Table by
default, which was causing a NULL pointer dereference. This was
happening very early in FDT parsing code, where our normal trap vector
is not set up yet, so it was jumping to the early_trap_vector, which
deadloops.

Added `no-plt` and `no-pic` flags to address this.

Fixes #43.