Do not increment mepc in fork()
mepc is incremented in syscall_epilogue, so an extra increment in fork()
was causing a jump to the wrong address. Incidentally, the userland
system call stubs contained an unnecessary push/pop of ra, which was
masking this bug.

The masking worked like this: normally, returning to mepc+4 would jump
to the next instruction after ecall, i.e. the "pop ra from stack". But
with the bug, it would skip the pop and jump straight to "release
stack". Since we set ra to the correct value via trap_frame mechanism,
overjumping the "pop ra" was not a problem, and thus, undetectable.

This also explains the concern expressed in the comment: why did it work
with the compressed instructions? Well, it jumped over two instructions,
straight onto 'ret', which also resulted in nothing worse than a tiny
one register-sized stack leak, undetectable with short-running programs.

The bug manifested itself loudly and proudly after removing the
superfluous stack operations. Now the excessive mepc increment would
jump over the 'ret' instruction, thus not returning from system call,
but immediately entering the next one (read() in our case).