Move pc to trap_frame and cleanup the way mepc is handled
syscall_epilogue used to increment mepc by 4 in order to step to the
next instruction. This is actually a wrong thing to do in general case.
We only want to increment mepc when it was a software interrupt, in
order to step over the `ecall` that calls into M-mode. In other cases we
want to return to the exact same instruction that we interrupted.

I discovered this with a user process spinning in a tight busy loop like
this:

    loop:
        j loop

that's a single instruction always jumping to itself. So if we step over
it, we break the loop and thus, break the user program.