Introduce cpu_t to hold the currently scheduled proc
Add a cpu_t structure to hold the state of a single core. Currently,
it's only a pointer to the currently scheduled proc (or null). Later,
both the state will grow, and the number of such slots will be equal to
the number of cores.

This makes proc_table.curr_proc index obsolete, just rely on cpu.proc
for all the same purposes.