Fix changes in timer frequencies
sifive_u machine uses 10x less frequent timer than the sifive_e and virt
machines, so split that into a separate header.

Not sure when that changed, but I was using a somewhat old qemu, and now
after switching to a new one, things started acting weird, which I
traced to this change in behavior.