Pipe fixes: better signalling and reference counting
Add a test program that does essentially 'cat file | wc -c'. This
exposed a number of problems with the generic case of working with
pipes, which this commit fixes.

Most significantly, don't use pids of writer/reader, invert the
dependency instead: the process now keeps a pointer to the object it's
waiting on. This way it's easy to avoid resurrecting the process that
already exited. Also, the same mechanism can be used later for waiting
on other objects for other reasons.