Fix stack overflow in ps and add a check for it
MAX_PROCS is only 8, so there's no point to allocate a buffer for 32
pids. And having 24 excessive ints amounts to 96 bytes, 20% of my entire
stack size!

This was causing a stack overflow in 64-bit build only (due to larger
stack allocations for registers).

Added a check for stack overflows in each syscall, hopefully that will
make it easier to debug next time.

This also suggests to make an attempt at allocating all kernel memory in
an area that's not accessible to the user processes, which increases the
chance to get an access violation exception in case of such overflows.