Rewrite shell's command traversal, pipes seem to work now
Rewrite traversal of the parsed command chain recursively and start
spawning processes from right to left. While at it, juggle the refcounts
of the pipe files more carefully so that none of the references get
leaked underefed.

Fix pipe_read() to check the buffer for data availability repeatedly
after a wakeup. We need that because the process can be woken up while
there's still nothing to read in the buffer, and we'd end up reading a
full buffer of garbage.