Initial implementation of LCD support
Add support for Hitachi HD44780 4-line LCD display. It's a complete
ripoff of the Arduino LiquidCrystal library, adapted to my interfaces
and simplified a bit by removing some generalities that I'm not
interested in.

HD44780 requires some time to process incoming commands and data.
Currently, this time is being ensured by spinning busy loops
conservatively overestimating the time that's really needed. This should
be improved by changing those busy loops to poll the is-busy pin on the
LCD itself instead of polling time.