Align stack to 16 bytes
Our stack section immediately follows bss, and thus can become aligned
at funny offsets depending on what global variables we have. A recent
attempt to introduce a struct with a sizeof a multiple of 2, but not 4
resulted in a misaligned stack that broke things in a way that was very
hard to debug (the system would mysteriously not boot at all on
HiFive1-revB, somewhere in the earliest stages, before kinit()).

Turns out, the ABI requires[1] sp to always be aligned to 16 bytes
(except on RV32E, where it must be aligned to 4).

Add an alignment directive to the linker script to keep the stack
independent of bss.

[1] https://github.com/riscv-non-isa/riscv-elf-psabi-doc