Fix deadlock on self in proc_pinfo()
This is a lingering problem with inconsistent process locks. My old
approach was to acquire a lock on own process in every syscall, but
later the scheduler was changed to acquire a lock on a process before
switch()ing to it. It’s been cleaned up here and there, but
inconsistently, some syscalls still acquire it. Moreover, a fresh
process starts running with lock=0, and the problem only becomes
apparent if the timer has a chance to run, which triggers the scheduler,
which in turn acquires the lock on the process before letting it
continue running.