Add kernel stack overflow protection
Set the lowest word of kstack_page to a sentinel magic number. If we
ever write a system call that overflows its stack, there's a great
chance it will ruin the sentinel. Check for such bad condition in
syscall() so that we can diagnose the problems sooner and more easily.

This exposed a hilarious off-by-one bug in proc_execv: it was storing
argv[] at the top of the process stack, but the terminating NULL of the
argv array used to step on the first word on the page *following* the
allocated user stack page. Sometimes that following page happens to be
another process' kstack_page, in which case we'd immediately overwrite
the magic sentinel.