Improve parent's ability to wait on children
Processes used to clean up immediately upon exit, which means that the
parent in wait() may actually miss some of their children exiting if the
scheduling doesn't go in a lucky order. Here's a way to reproduce it:

    cat file | wc missing-file

wc would try to read a missing-file and exit with ENOENT, closing the
pipe. Since the pipe closes the reading end, cat would also exit
immediately as there's no point in writing to a broken pipe. All of that
would happen before the shell has a single chance to get scheduled, and
thus, it would remain sleeping, as there's nothing to wake it up.

This introduces the famous Zombie state for the process. An exiting
child will set its state to Zombie and only do a partial cleanup of
itself. Then, in proc_wait(), the parent will look for zombie children
of itself, and only then complete their cleanup.