Do a better job at cleaning up running qemus
Since we do `docker run` here, the pid we get from Popen.pid is for
`docker`, not for the container that actually runs qemu, and thus, when
we try to p.terminate(), it's not affecting the container. So try to
kill the actual container when we want to quit, not the docker cli.