Prevent switch statement conversion to a jump table
Set a threshold to convert the chain of branches into a jump table when
compiling the switch statement. We have recently seen[1] another case of
it, but that was trivial code, easy to convert, and today it's a large
complicated switch statement in printf() implementation.

This workaround can be removed when proper user executables are
implemented with their own .rodata sections that they have access to.

[1]: 3410fa4434d3daaa345161cd6ead1f1a13325cb5