Prepare privileged status registers for easy switch to S-mode
Hardcoding M-mode status registers is not good enough, as they will be
not accessible in S-mode. So we need an easy switch between the two
sets.

mhartid should be the only use case that is left. We need to get rid of
it altogether, since it's completely inaccessible from S-mode.