Get rid of (almost) all direct access to mhartid
When we boot in S-mode, a0 will be set to hartid by the bootloader. When
we boot in M-mode, do the same, and later use a0.

The only couple places where we try to access hartid at runtime were
hardcoded to 0 for now. Later we'll need to store hartid in a trap_frame
or some similar place if we need to access it in interrupt handlers.