Implement set_timer_after() on Ox64
On Ox64 we don't have access to CLINT timer like we do elsewhere. On the
other hand, we're running under SBI, so we can issue a "syscall" into
M-mode to ask to set the timer for us.

sbi_ecall() is copied from Linux almost verbatim, but there are very few
ways how it can be spelled anyway.