Fix handling of user mode
When we return from a system call, we typically think that it's the same
software interrupt trap we're returning from. However, a process may go
to sleep during a syscall, and the next time it wakes, it may find
itself being woken up from within a hardware interrupt, which was
trapped in M- or S-mode. So we need to properly ensure that we will be
exiting the system call into U-mode, not back into something privileged.