Fix Ox64 RAM start address
This was a nasty mistake. The bootloader relocated the binary to the
right place, so most of it was functioning correctly, except for things
where we have addresses stored in .data segment, like userland program
entry points and syscall vector. Those would remain pointing to the
wrong addresses. Figuring this out was a pain without gdb.