Fix printf of negative numbers
There were a couple problems: the parameter for print_radix was coerced
to an unsigned number, and the buffer holding the printed number wasn't
large enough to hold a very large int64.

It would be nice to avoid using a terminating zero for the temporary
values within printf(), but that's a larger surgery than I want to
undertake now.