Rearrange PMP code, make room for disabling pages on demand
Before this, we had four memory ranges covered by PMP, the last one of
them covered the entire heap that's dedicated to paged allocations.
Since PMP ranges are effective in-order, this meant that all subsequent
changes to PMP would never take effect.

Thus, move that last section to reside in the uppermost PMP entry. This
leaves the mid-range of PMP configs available for disabling access
rights dynamically, which will be useful in tests for detecting memory
over/underflows.