Fix 64-bit sifive_e target, as much as possible
Looks like the 64-bit sifive_e is also hardcoded in qemu to have 16k
RAM. With this limit, the kernel no longer hangs, but execv() fails,
because not enough RAM remains after bss for allocation.

Added a note to the readme about that. It feels like it should be
possible to streamline some things and make it fit in the available
memory - some day.