Implement virtual memory
This is an early attempt, and rather hacky. All the memory that the
kernel is supposed to reach (i.e. the code, data, and memory-mapped
registers) are identity-mapped in init_vpt().

All the userland code and data is mapped by stripping everything above
the 9 bits of vpn0. In addition to that, the user VM pagetable contains
a kernel-executable superpage mapping the entire RAM with identity
mapping, just like the kernel pagetable. That's there to allow the
kernel to keep executing after it sets the user satp. Otherwise the very
next instruction would fail to be fetched.

A lot of code should be generalized after this initial cut. E.g. the
UVA/UPA macros that map the user pointers to virtual address and to
physical address are a complete hacks. map_page(), map_l2page() and
map_superpage() are overly rigid as well.