Enable virtual memory on qemu virt
This implements the missing peaces and enables paged virtual memory on
the only machine where it works at the moment, the qemu virt.

In summary, this keeps track of the user page table, sets satp to
appropriate user-or-kernel value, and converts the pointers to and from
the virtual address space as needed. Only userland pointers need the
conversion, as the kernel operates with identity mapping, so the virtual
and physical addresses always match.

PAGE_SIZE is increased to 4 KiB because Sv39 does not allow a finer
granularity.