Simplify syscall_dispatch()
Similarly to kernel_timer_tick(), pass sp from the trap handler to C
code so that it can save it in the process's kernel-side context. Saving
all other registers doesn't seem to be necessary at this point.

The code being removed was introduced with 4ba8eca9, which explicitly
aimed at solving a stack leak.