Fix excessive stack usage in userland printf()
Implementation of printf() allocates a 21-byte buffer on the stack to
have a place to format numbers to. This was observed to cause a stack
overflow in some otherwise quite reasonable cases (our stack is
typically only 512 bytes on most targets).

Now that we have a guard page below the stack, and an option to
artificially force the stack to be tiny, create a test that guards
against regressions in printf(). The userland program 'testprintf'
allocates a sizeable placeholder buffer on the stack and calls printf().
Without a fix, this reproduces the stack overflow.

This commit contains a fix as well: instead of allocating a separate
buffer for numbers on the stack, reserve the end of the general buffer
for that purpose. The general buffer is typically the entire page, so
that should be sufficient for both purposes.

kprintf() prints directly to UART, so there's no general buffer to
reuse, so it still allocates the number buffer on the stack. But that
should be less of a problem, as we mostly use kprintf() for debugging,
it's not really called in the regular kernel codepaths.

Interestingly, this commit exposes some unknown stack overflow bug
somewhere in the smoke test, which remains to be investigated.