Make debug symbols available in userland address space
When working with virtual memory, userland code is executed at a
different address than specified in the binary. This means that when gdb
finds itself in the user code, it doesn't know it has symbols for it,
but for another address range.

This change loads the symbols for the second time, but with an offset.
The trickiest part of it was to figure out that a negative offset has to
be specified as two's complement:

    add-symbol-file out/os_virt -o 0xffffffff80000000

This approach has a downside: all symbols will now have two locations:

    (gdb) b u_main_shell
    Breakpoint 2 at 0xaa2c: u_main_shell. (2 locations)
    (gdb) info breakpoints
    Num     Type           Disp Enb Address            What
    2       breakpoint     keep y   <MULTIPLE>
    2.1                         y   0x000000000000aa2c in u_main_shell at user/src/shell.c:223
    2.2                         y   0x000000008000aa2c in u_main_shell at user/src/shell.c:223

That shouldn't be a problem, though, as the kernel- and user-code spaces
are distinct and do not overlap.

It may prove to be a bit treacherous, though, as gdb may choose to print
a deceptive message where it placed the breakpoint that will actually
get triggered. In this example it says the breakpoint is at 0x180, but
actually the other one at 0x80000180 will get triggered:

    (gdb) b trap_vector
    Breakpoint 3 at 0x180: trap_vector. (2 locations)
    (gdb) info breakpoints
    [...]
    3.1                         y   0x0000000000000180 src/boot.S:250
    3.2                         y   0x0000000080000180 src/boot.S:250

`disassemble-next-line auto` may come in handy if something goes bonk
with symbols, it will disassemble the next instruction in such case.