Add a detach() systeam call
The system call unlinks the current process from its parent and wakes up
the parent just in case it was sleeping. This in effect daemonizes the
current process, leaving it to run in the background and giving the
control back to shell.

Currently, the behavior of the daemon can only be observed with a
debugger, but some simple IPC mechanism will allow us to talk to the
daemon and inquire its state.

This required a fix in the scheduler. Now it calls copy_trap_frame()
unconditionally whenever it switches to a target process. Before now, it
was only necessary when waking up from a timed sleep.