Parameterize test script argument
Pass the path to the script file via bootargs. This will make it easier
to add more scripts. Rename "smoke-test" to "test-script" to better
reflect that.

As a side effect, this gets rid of one extra process when running tests
(the "smoke-test" program), which in turn reduces the memory footprint
(2-7 pages, depending on the configuration).

The path to the test script is passed to the shell program via argv.
Shell's stack is used to store the string and the argv array itself,
which consumes some stack (quite significant in case of tiny-stack
test), which required a supporting change in run_shell_script(). We no
longer allocate `pbuf` on the stack, but reserve a bit of the file
buffer for it. This saving is enough to compensate for the expenses for
argv.

The shell now has an optional parameter '-f', which makes it freeze
after executing the script. This is needed to avoid the top level shell
from exiting.