Add test exercising the Fibonacci Daemon
Add a test to launch the `fibd` program and ask it for a response a
couple times.

Due to differences in timing (and the test's sensitivity to that), the
result gives off-by-one results on sifive_u machine, so keep it
commented out for now. A more reliable mechanism would help.