Fix sp in early _start_kernel for non-boot harts
This is one of those cases:

    I really don't see how this could have gotten broken...
    After investigation, I really don't see how this ever worked.

There's a piece of memory dedicated for stack of booting harts, between
stack_bottom and stack_top (defined in kernel.ld). Due to ambiguity in
stack growth direction being the opposite of address growth direction,
there was a mistake in assigning a piece of that memory for stack use.

Since the boot hart is hart0, stack_top+0 is the same as stack_top-0 and
the booting hart was assigned the correct memory range for its stack.
That was not the case for hart1 though, it was assigned stack_top+512
instead of the correct stack_top-512. And thus, for a short period of
time that hart1 was working before being put to sleep, it was using the
memory beyond stack_top for its stack purposes.

That memory is used for allocatable paged memory, so in essence hart1
was thrashing occasional word or a few in paged_memory.pages[0], which,
incidentally, is kalloc()ed for userland stack of the first process (the
shell).

Typically, the shell would start with an empty stack so someone
thrashing the memory there wasn't detectable; it was treated as
rewritable uninitialized memory anyway. However, a recent change[1]
started using the very top of stack page for passing argv into the shell
process when running scripts (as is the case for the automated tests)
and it became detectable.

The problem was still probabilistic: some 30%-50% of times the tests
would pass. When hart1 raced ahead and thrashed the memory before hart0
went on with the full init, the tests would pass because hart0 would
write argv on top of the hart1's garbage. But when hart0 was faster to
acquire init_lock, it would go on initializing argv, and hart1 would
only get a chance to do its dirty work after that, overwriting bits of
argv.

No, it wasn't easy to diagnose :-)

Learned a couple things about gdb on the way:

1. Watchpoints set with `watch <address>` were thread-local by default
   on my configuration, as they defaulted to hardware watchpoints. This
   was mighty confusing as they were triggered when hart0 touched the
   memory, but then the memory location would "magically" change the
   value without triggering the watchpoint. A way to get around that was
   to `set can-use-hw-watchpoints 0`. This forces gdb to use software
   watchpoints, which are much, MUCH slower, but get things done:
   modification by both harts would now trigger the watchpoint.
2. Since my `add-inferior` command didn't specify the binary to use, the
   second inferior didn't see any symbols, and breakpoints would only be
   set to the first inferior (corresponding to hart0). So lesson
   learned, pay attention to what `info breakpoints` says:

   Wrong (notice all breakpoints having the "inf 1" suffix, even when I
   explicitly asked it to "stop only in inferior 2", which is clearly a
   "never trigger" condition):

        (gdb) info breakpoints
        Num     Type           Disp Enb Address            What
        ...
        3       breakpoint     keep y   0x0000000080009a38 in u_main_shell at user/src/shell.c:223 inf 1
        4       breakpoint     keep y   0x0000000080000040 src/boot.S:44 inf 1
                stop only in inferior 1
        5       breakpoint     keep y   0x0000000080000040 src/boot.S:44 inf 1
                stop only in inferior 2

   Right (notice some breakpoints having the "inf 2" suffix):

        (gdb) info breakpoints
        Num     Type           Disp Enb Address            What
        ...
        3       breakpoint     keep y   <MULTIPLE>
        3.1                         y   0x0000000080009a38 in u_main_shell at user/src/shell.c:223 inf 1
        3.2                         y   0x0000000080009a38 in u_main_shell at user/src/shell.c:223 inf 2
        4       breakpoint     keep y   <MULTIPLE>
                stop only in inferior 1
        4.1                         y   0x0000000080000040 src/boot.S:44 inf 1
        4.2                         y   0x0000000080000040 src/boot.S:44 inf 2
        5       breakpoint     keep y   <MULTIPLE>
                stop only in inferior 2
        5.1                         y   0x0000000080000040 src/boot.S:44 inf 1
        5.2                         y   0x0000000080000040 src/boot.S:44 inf 2

This whole story likely also explains mysterious failures when I tried
to let the hart1 be the one that boots instead of hart0. I didn't want
to spend time investigating back then, but now it seems to have sorted
itself out. If this proves to be true, maybe sifive_u target can now be
switched to the U-core and get itself an MMU implementation.

[1]: https://github.com/rtfb/riscv-hobby-os/commit/316a91e4d4d7058bbcc37119edee470b402343f0