Fix a bug in jumping over hart sync code
The code was comparing 0 == 1 and never took that branch. This doesn't
seem to have hurt anything since the hart sync code seems to work well
in the single-harted case, but for clarity, let's have it fixed.

This means this check can be eliminated altogether, but it feels simpler
to not go through the sync dance in the single-harted case.