Extend wait() syscall with more elaborate wait conditions
wait() was only able to wait for the child process to exit. Now it
(optionally) takes a parameter describing the condition for wakeup.

If the parameter is null, it behaves like before - waits for any child
to exit. Otherwise, the structure allows specifying the type of
condition and parameters, if needed.

The new wait type is `WAIT_COND_NSCHEDS` - wait for a target process
(specified by `.target_pid`) to be scheduled exactly N times (specified
by `.want_nscheds`). This allows scheduling waits independent of exact
timing.

`WAIT_COND_CHILD` implements the old behavior, equivalent to specifying
null.

On the kernel side, the parameter is converted to another structure,
that is identical for now, but may change independently of the
user-facing struct. That's because of two reasons:

1. The naming that makes most sense for the user does not necessarily
   makes most sense on the kernel side. So some constants are renamed
   (e.g. WAIT_COND_CHILD -> PWAKE_COND_CHAN).
2. Some kinds of conditions are not exposed to the user. For example,
   the process may go to sleep blocked on a pipe. This separation allows
   more flexibility and extensibility on the kinds of conditions.