Implement /proc/<pid>/stats file
Each process now has a `stats` file under /proc and it contains a few
pieces of data:
* parent pid (or -1 if the process is parentless)
* number of times the process was scheduled
* number of pages allocated in the name of the process

The stats file is added to the bifs filesystem upon creation of each
process. The file sits dataless until someone opens the file. Upon
opening, a page of memory is allocated and the content is sprintf()'ed
into that page. Afterwards, reading from that file can proceed like from
the regular file. When the file is closed, the underlying page with data
is released. The next time the file gets opened, the new backing memory
is allocated and an updated content is sprintf()'ed.