Initial port to Star64
Everything seems to be working quite well, most importantly, including
the MMU. This is the first hardware that I managed to get the MMU
working (and there wasn't any adjustments to be made - the last fixes to
make it work on sifive_u were sufficient).

There are a few files that are complete or almost-complete copies of
other files, that remains to be cleaned up (marked with TODOs).

The boot address behaves a bit differently to what I observe on other
machines. I compile the binary to be located at the address 0x80200000,
but in the binary header I specify 0x40200000 - the bootloader seems to
automatically add 0x40000000 to it (the DRAM start address, according to
the docs). Then in the U-Boot, I specify 0x80200000 as well.

The trailing "- fffc6aa0" at the U-Boot booti command specifies the
address of the builtin device tree file. It refuses to boot without it.