Add clock program: count seconds visually
The program increments a seconds counter to visually measure my timing
routines against wall clock.

As is, it's running too slow because my kernel timer interrupt only runs
once a second, so it's possible that the scheduler won't even know it's
time to wake up a process for a significant fraction of a second. So a
real test should be performed with adjusted KERNEL_SCHEDULER_TICK_TIME.

Even with the timer adjusted, it still demonstrates my frequencies are
somewhat off on some targets. Fixed the Star64 frequency, as it was
seriously off.