Speedup response to user input
There was a delay between the time the user presses Enter and the time
when the scheduler has a chance to run. That's because the timer is now
deliberately set to a very infrequent 1Hz pace, and the sleeping
scheduler didn't have a chance to know that the UART driver has marked a
process for wakeup.

This commit implements a way for the sleeping scheduler to wake up. It
is achieved by setting a global variable unsleep_scheduler to true. When
the sleeping scheduler (now implemented in soft_park_hart()) encounters
an unsleep, it induces a timer interrupt as soon as possible, which
triggers the usual scheduling workflow.

It may seem like an overkill to induce the timer interrupt when we could
simply return from the sleeper back into the scheduler, but we can't do
that. The scheduler and the interrupts use the same stack (see
set_global_stack) and interrupts corrupt the frame unwind pointers
leading to soft_park_hart(). Using separate stacks for these purposes
would help, but that would bloat the footprint on the smallest supported
systems.