Move leaky tests to a separate script
The smoke test used to run a special program that detaches itself from
the shell and hangs there forever. This is intentional test that
validates that the resources get leaked in an expected and controlled
way. However, this interferes with running other tests on tiny targets
as it wastes precious memory.

Move that to a separate script so that the main smoke test could run
more good things.