Speed up tests
Don't wait for a timeout, when the test is done, signal to qemu-launcher
that it's done. This is achieved by printing the string "QUIT_QEMU" to
stdout.

The "dry-run" test does nothing in the userland, so it needs to still
rely on the timeout, but the timeout doesn't need to be that long, 1s is
enough.

Some tests still run suspiciously long: daemon test should finish
faster, this needs investigation.

The complete test run got reduced from more than 1m to less than 25s.