Reduce MAX_PROCS to 4 on sifive_e targets
sifive_e targets are sooo small that the kernel fails to even start on
the 64-bit sifive_e. So start reducing the bss segment where possible.

This reduces the proc_table footprint and allows the kernel to run the
shell. But the shell still can't spawn any children, so need to trim
some more.