Improve shell error reporting
Print errno in addition to status code when exec() fails. Also, when
exec() fails due to OOM, it can't allocate the memory for printing
needs, so it would fail silently in that case. To avoid complete
silence, print a constant "OOM" message in that case.

This has increased consumption of stack of traverse() by 16 bytes, so
needed a complementary fix to avoid a stack overflow on tiny stacks. The
fix avoids a terminating recursive call. Instead, it rearranges the
check for termination to be performed in one call above it.