Fix a bug in fs_free_file()
Failure to zero out tmpfile_mem leaves the file entry in a broken state.
Subsequent use of that file entry can cause problems. For example,
reading any procfs file (which allocates tmpfile_mem) and after that
doing anything with pipes would cause corruption of the file object
because it would erroneously take this branch again and corrupt the pipe
structure.