Reduce stack on sifive_e, enable remaining tests
sifive_e is known to always be single-core so 4KiB of preallocated
boot-time stack is too much. And that's like HALF of the remaining
memory. So reduce that to 512 bytes (on 32-bit) and 1KiB (on 64-bit).

It's not yet clear why the 64-bit build fails with 512 bytes, it should
be enough there as well. Will investigate that later.

For now, this allows enabling the remaining tests on the 64-bit sifive_e
as well.