Generate all syscall glue code
A lot of syscalls glue code is very repetitive and adheres to the same
pattern. It's easily generatable from a very simple and concise source
of truth. That source of truth is now in include/syscalls.hh, and the
scripts/gen-syscalls.py generates the following files from it:
* include/syscallnums.h
* include/syscalldecls.h
* src/syscalls.c
* user/inc/usyscalls.h
* user/src/usyscalls.S

All these files are still tracked in git, both for accountability, and
for easier readability of complete code.

A new syscall should now be added only to include/syscalls.hh, the rest
will get dealt with by the Makefile.