Fix stack overflow in init code
Turns out my kernel init code overflows stack. It wasn't easily
detectable because there was enough unused memory under the stack. I
discovered that when I tried to enable SMP and allowed another hart to
use the stack that's right under the one used by the boot hart.

The fix is big and fiddly, but it's all about reducing the stack
footprint. This is achieved by three kinds of fixes:
1. Rewrite a function into a macro
2. Remove local variables where possible
3. Rearrange code to eliminate a function call